GVT-1597: Move publication stuff under one folder
