+++ ui/src/i18n/config.ts
@@ -16,7 +16,7 @@ import kmPostDialogFi from 'tool-panel/km-post/dialog/translations.fi.json';
 import trackNumberDialogFi from 'tool-panel/track-number/dialog/translations.fi.json';
 import linkingFi from 'linking/translations.fi.json';
 import userFi from 'user/translations.fi.json';
-import publicationLogFi from 'publication-log/translations.fi.json';
+import publicationLogFi from 'publication/log/translations.fi.json';
 
 export const resources = {
     en: {
