+++ ui/src/publication/card/publication-card.scss
@@ -1,5 +1,5 @@
-@import '../vayla-design-lib/typography';
-@import '../vayla-design-lib/colors';
+@import '../../vayla-design-lib/typography';
+@import '../../vayla-design-lib/colors';
 
 $color-error: $color-red-dark;
