+++ ui/src/publication/card/publication-list.scss
@@ -1,5 +1,5 @@
-@import '../vayla-design-lib/typography';
-@import '../vayla-design-lib/colors';
+@import '../../vayla-design-lib/typography';
+@import '../../vayla-design-lib/colors';
 
 .publication-list__header {
     margin-bottom: 12px;
