+++ ui/src/publication/card/publication-list.tsx
