+++ ui/src/publication/log/publication-log-link.tsx
