+++ ui/src/publication/log/publication-log-table-entry-mappings.ts
