+++ ui/src/publication/log/publication-log.scss
@@ -1,5 +1,5 @@
-@import '../vayla-design-lib/colors';
-@import '../vayla-design-lib/typography';
+@import '../../vayla-design-lib/colors';
+@import '../../vayla-design-lib/typography';
 
 .publication-log-table__container {
     overflow-y: auto;
