+++ ui/src/publication/table/publication-table-item.scss
@@ -1,6 +1,6 @@
-@import '../vayla-design-lib/colors';
-@import '../vayla-design-lib/shadows';
-@import '../vayla-design-lib/typography';
+@import '../../vayla-design-lib/colors';
+@import '../../vayla-design-lib/shadows';
+@import '../../vayla-design-lib/typography';
 
 $color-ok: $color-green-dark;
 $color-error: $color-red-dark;
