+++ ui/src/publication/table/publication-table-item.tsx
