+++ ui/src/publication/table/publication-table.scss
